+++ lib/connection/pool.js
@@ -8,7 +8,7 @@ var inherits = require('util').inherits,
   f = require('util').format,
   Query = require('./commands').Query,
   CommandResult = require('./command_result'),
-  assign = require('../topologies/shared').assign;
+  assign = require('../utils').assign;
 
 var MongoCR = require('../auth/mongocr')
   , X509 = require('../auth/x509')
@@ -618,7 +618,7 @@ Pool.prototype.connect = function() {
       moveConnectionBetween(connection, self.connectingConnections, self.availableConnections);
 
       // Emit the connect event
-      return self.emit('connect', self);      
+      return self.emit('connect', self);
     }
 
     // Apply any store credentials
@@ -868,7 +868,7 @@ Pool.prototype.destroy = function(force) {
       }
     }
 
-    // Destroy the topology 
+    // Destroy the topology
     return destroy(self, connections);
   }
 
@@ -905,7 +905,7 @@ Pool.prototype.destroy = function(force) {
 
       destroy(self, connections);
     // } else if (self.queue.length > 0 && !this.reconnectId) {
-      
+
     } else {
       // Ensure we empty the queue
       _execute(self)();
