+++ lib/topologies/mongos.js
@@ -8,7 +8,7 @@ var inherits = require('util').inherits,
   retrieveBSON = require('../connection/utils').retrieveBSON,
   MongoError = require('../error'),
   Server = require('./server'),
-  assign = require('./shared').assign,
+  assign = require('../utils').assign,
   clone = require('./shared').clone,
   sdam = require('./shared'),
   diff = require('./shared').diff,
@@ -268,7 +268,7 @@ function handleEvent(self) {
     self.emit('serverClosed', {
       topologyId: self.id,
       address: this.name
-    });     
+    });
   }
 }
 
@@ -393,7 +393,7 @@ function connectProxies(self, servers) {
       self.emit('serverOpening', {
         topologyId: self.id,
         address: server.name
-      });     
+      });
 
       // Emit the initial topology
       emitTopologyDescriptionChanged(self);
@@ -560,7 +560,7 @@ function reconnectProxies(self, proxies, callback) {
       self.emit('serverOpening', {
         topologyId: server.s.topologyId != -1 ? server.s.topologyId : self.id,
         address: server.name
-      });     
+      });
 
       // Add temp handlers
       server.once('connect', _handleEvent(self, 'connect'));
@@ -766,7 +766,7 @@ Mongos.prototype.destroy = function(options) {
     self.emit('serverClosed', {
       topologyId: self.id,
       address: x.name
-    });             
+    });
 
     // Destroy the server
     x.destroy(options);
