+++ lib/topologies/replset.js
@@ -10,7 +10,7 @@ var inherits = require('util').inherits,
   MongoError = require('../error'),
   Server = require('./server'),
   ReplSetState = require('./replset_state'),
-  assign = require('./shared').assign,
+  assign = require('../utils').assign,
   clone = require('./shared').clone,
   Timeout = require('./shared').Timeout,
   Interval = require('./shared').Interval,
@@ -435,7 +435,7 @@ var pingServer = function(self, server, cb) {
 
 // Each server is monitored in parallel in their own timeout loop
 var monitorServer = function(host, self, options) {
-  // If this is not the initial scan  
+  // If this is not the initial scan
   // Is this server already being monitoried, then skip monitoring
   if(!options.haInterval) {
     for(var i = 0; i < self.intervalIds.length; i++) {
@@ -470,7 +470,7 @@ var monitorServer = function(host, self, options) {
         }
 
         // Filter out all called intervaliIds
-        self.intervalIds = self.intervalIds.filter(function(intervalId) { 
+        self.intervalIds = self.intervalIds.filter(function(intervalId) {
           return intervalId.isRunning();
         } );
 
@@ -550,7 +550,7 @@ function topologyMonitor(self, options) {
 
   if(_process === Timeout) {
     return connectNewServers(self, self.s.replicaSetState.unknownServers, function(err) {
-      // Don't emit errors if the connection was already 
+      // Don't emit errors if the connection was already
       if(self.state === DESTROYED || self.state === UNREFERENCED) {
         return;
       }
@@ -583,7 +583,7 @@ function topologyMonitor(self, options) {
       }
 
       connectNewServers(self, self.s.replicaSetState.unknownServers, function() {
-        var monitoringFrequencey = self.s.replicaSetState.hasPrimary() 
+        var monitoringFrequencey = self.s.replicaSetState.hasPrimary()
           ? _haInterval : self.s.minHeartbeatFrequencyMS;
 
         // Create a timeout
