+++ lib/topologies/server.js
@@ -15,7 +15,7 @@ var inherits = require('util').inherits,
   ThreeTwoWireProtocolSupport = require('../wireprotocol/3_2_support'),
   BasicCursor = require('../cursor'),
   sdam = require('./shared'),
-  assign = require('./shared').assign,
+  assign = require('../utils').assign,
   createClientInfo = require('./shared').createClientInfo;
 
 // Used for filtering out fields for loggin
