+++ lib/topologies/shared.js
@@ -180,31 +180,6 @@ var inquireServerState = function(self) {
   };
 }
 
-// Object.assign method or polyfille
-var assign = Object.assign ? Object.assign : function assign(target) {
-  if (target === undefined || target === null) {
-    throw new TypeError('Cannot convert first argument to object');
-  }
-
-  var to = Object(target);
-  for (var i = 1; i < arguments.length; i++) {
-    var nextSource = arguments[i];
-    if (nextSource === undefined || nextSource === null) {
-      continue;
-    }
-
-    var keysArray = Object.keys(Object(nextSource));
-    for (var nextIndex = 0, len = keysArray.length; nextIndex < len; nextIndex++) {
-      var nextKey = keysArray[nextIndex];
-      var desc = Object.getOwnPropertyDescriptor(nextSource, nextKey);
-      if (desc !== undefined && desc.enumerable) {
-        to[nextKey] = nextSource[nextKey];
-      }
-    }
-  }
-  return to;
-}
-
 //
 // Clone the options
 var cloneOptions = function(options) {
@@ -234,7 +209,7 @@ function Interval(fn, time) {
 
   this.isRunning = function () {
     return timer !== false;
-  };  
+  };
 }
 
 function Timeout(fn, time) {
@@ -256,7 +231,7 @@ function Timeout(fn, time) {
   this.isRunning = function () {
     if(timer && timer._called) return false;
     return timer !== false;
-  };  
+  };
 }
 
 function diff(previous, current) {
@@ -298,7 +273,7 @@ function diff(previous, current) {
 
     // Go over all the previous servers
     for(var i = 0; i < previous.servers.length; i++) {
-      if(previous.servers[i].address.toLowerCase() 
+      if(previous.servers[i].address.toLowerCase()
         === current.servers[j].address.toLowerCase()) {
         found = true;
         break;
@@ -346,7 +321,6 @@ module.exports.getTopologyType = getTopologyType;
 module.exports.emitServerDescriptionChanged = emitServerDescriptionChanged;
 module.exports.emitTopologyDescriptionChanged = emitTopologyDescriptionChanged;
 module.exports.cloneOptions = cloneOptions;
-module.exports.assign = assign;
 module.exports.createClientInfo = createClientInfo;
 module.exports.clone = clone;
 module.exports.diff = diff;
